no poder nivel si esta asociado a un curso

+++ src/services/nivelesService.js
@@ -1,4 +1,5 @@
 const Nivel = require('../models/nivel');
+const Curso = require('../models/curso');
 
 exports.getAllNiveles = async () => {
   return await Nivel.find();
@@ -18,6 +19,14 @@ exports.updateNivel = async (id, data) => {
 };
 
 exports.deleteNivel = async (id) => {
+  // Verificar si existen cursos que referencian este nivel
+  const cursosAsociados = await Curso.countDocuments({ id_nivel: id });
+  if (cursosAsociados > 0) {
+    const err = new Error(`No se puede eliminar el nivel porque existen ${cursosAsociados} curso(s) asociados.`);
+    err.statusCode = 400; // Bad Request / Regla de negocio
+    throw err;
+  }
+
   const result = await Nivel.findByIdAndDelete(id);
   return !!result;
 };
